Reworked the view of the cell a bit
