+++ Eye in the sky/Eye_in_the_skyApp.swift
@@ -10,8 +10,8 @@ import SwiftUI
 @main
 struct Eye_in_the_skyApp: App {
     var body: some Scene {
-        WindowGroup<ContentView> {
-            ContentView()
+        WindowGroup<WeatherListView> {
+            WeatherListView()
         }
     }
 }
